rag-demos/app_host/Python.Plugins/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
fastapi==0.115.2
idna==3.10
pip==24.2
pydantic_core==2.23.4
pydantic==2.9.2
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
pandas
yfinance[nospam,repair]
plotly-express
azure-identity
azure-storage-blob
async_timeout
xmltodict
numpy
requests
uvicorn
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai
openai